+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * CORE: [FIX] Open selective halts sync #174
 * CORE: [FIX] Avoid concurrent refreshAccessToken requests #176
+* CORE: [FIX] Avoid creating login item for app translocation paths #177
 
 
 ## 1.0.0-rc1